feat(to semver): To Semver helper function broken out into its own require ready file.
